+++ playground/backend/commands/playground.ts
@@ -1,6 +1,7 @@
 import { BaseCommand } from '@adonisjs/core/ace'
 import type { CommandOptions } from '@adonisjs/core/types/ace'
-import mail from '@adonisjs/mail/services/main'
+import vine from '@vinejs/vine'
+import { z } from 'zod'
 
 export default class Playground extends BaseCommand {
   static commandName = 'playground'
@@ -13,10 +14,29 @@ export default class Playground extends BaseCommand {
   async run() {
     this.logger.info('Welcome to playground!')
 
-    await mail.send((message) => {
-      message.subject('Test mail')
-      message.to('servicexskull@gmail.com')
-      message.text('Test mail')
+    const res = vine.validate({
+      schema: vine.object({
+        user: vine.zod(
+          z.object({
+            name: z.string().min(2).max(10).describe('Name of the user'),
+            age: z.number().positive().min(18).max(100).describe('Age of the user'),
+          })
+        ),
+        goat: vine.object({
+          name: vine.string().minLength(5).maxLength(100),
+          age: vine.number().positive().min(18).max(100),
+        }),
+      }),
+      data: {
+        user: {
+          name: 's',
+        },
+        goat: {
+          name: 'k',
+        },
+      },
     })
+
+    console.log(res)
   }
 }