+++ packages/castle/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @folie/castle
 
+## 0.0.28
+
+### Patch Changes
+
+- refactoring in vine-zod
+
 ## 0.0.27
 
 ### Patch Changes